Ignore DoesNotExist and ErrPermission in recursive ReadDir
These can occur during recursion, but are not necessarily a problem of
the ReadDir, so just assume such problematic files are not directories
and return them. The caller can then figure out whether they are even
needed at all, and an error should be returned. This allows simply
ignoring a broken symlink when it would not even be used because of its
extension.